Allow elements to call services
I added a default case to the onElementClick to allow calling of 'normal' HASS services with a svg element. I use this for an alarm numpad. But it can also be used to have multiple elements triggering the same entity_id (e.g. scenes). I also added a check for the second elements to prevent errors for the hidden layer handling.

If there is a plan to include further actions which are floorplan exclusive this is still possible by adding more cases to the switch statement.

~Cheers

PS: I hope the format is okay I think I did use auto format in between edits in my IDE